(PC-29977)[PRO] refactor: Deleting GetCollectiveOfferFormDataApdater Adapter.
